Implement FSGROUP support correctly
An FSGROUP can have a single member.  We define an FSGROUP response
as something for which all returned records have an integer >= 1 as
the last field.

Technically, the definition of an FSGROUP is "the definition of any
given filesys type which has one additional final field", but that
requires knowing about each filsys type, which is liblocker's job.